split the explore genre and created the Genre card
