+++ frontend/components/home/ExploreGenres.tsx
@@ -6,7 +6,7 @@ import comedyImage from "../../styles/images/genres/comedy.png";
 import politicsImage from "../../styles/images/genres/politics.png";
 import crimeImage from "../../styles/images/genres/crime.png";
 import otherImage from "../../styles/images/genres/other.jpg";
-import NexLink from "next/link";
+import GenreCard from "../cards/GenreCard"
 
 // Define the genres array with name, image, and link properties
 const genres = [
@@ -18,71 +18,11 @@ const genres = [
   { name: "Other", image: otherImage, link: "Other" },
 ];
 
-const GenreCard = ({ genre, isHovered, onMouseEnter, onMouseLeave }) => {
-  const scale = isHovered ? "scale(1.1)" : "scale(0.95)";
-
-  return (
-    <NexLink href={`/Explore/Genre/${genre.link}`} passHref>
-      <Box
-        key={genre.name}
-        h="100%"
-        borderRadius="1.2em"
-        overflow="hidden"
-        position="relative"
-        _hover={{
-          boxShadow: "xl",
-          cursor: "pointer",
-          transition: "all 0.4s ease-in-out",
-        }}
-        transform={scale}
-        transition="all 0.4s ease-in-out"
-        boxSizing="border-box"
-        outline={"2px solid rgba(255, 255, 255, .3)"}
-        onMouseEnter={() => onMouseEnter(genre.name)}
-        onMouseLeave={onMouseLeave}
-        zIndex={isHovered ? 1 : 0}
-      >
-        <Image
-          src={genre.image.src}
-          alt={`${genre.name} background`}
-          width="100%"
-          height="100%"
-          objectFit="cover"
-          opacity="0.8"
-        />
-        <Text
-          fontWeight="bold"
-          fontSize={["md", "lg", "xl"]}
-          position="absolute"
-          left="50%"
-          top="50%"
-          transform="translate(-50%, -50%)"
-          textShadow="2px 2px 4px #000000"
-          p={1}
-        >
-          {genre.name}
-        </Text>
-      </Box>
-    </NexLink>
-  );
-};
-
 const ExploreGenres = () => {
   const [hoveredGenre, setHoveredGenre] = useState(null);
 
   return (
     <Box marginBottom="3em">
-      <Box
-        w={{ base: "70%", md: "20%" }}
-        top={0}
-        left={0}
-        zIndex={-99}
-        borderRadius={"0.5em"}
-      >
-        <Text fontSize="2xl" fontWeight="bold" mb={"1em"} ml={"0.7em"}>
-          Explore Genres
-        </Text>
-      </Box>
       <Grid
         templateColumns={["repeat(2, 1fr)", "repeat(3, 1fr)", "repeat(6, 1fr)"]}
         gap={4}
